Move to ruby-lsp-rails native addon
