+++ lib/ruby_lsp/rails/factory_bot/addon_name.rb
@@ -1,0 +1,7 @@
+module RubyLsp
+  module Rails
+    module FactoryBot
+      ADDON_NAME = "ruby-lsp-rails-factory-bot".freeze
+    end
+  end
+end
